new module name + extension function API
