+++ sdk/forms/src/main/AndroidManifest.xml
