+++ sdk/forms/src/main/assets/InAppFormsTemplate.html
