+++ sdk/forms/src/main/java/com/klaviyo/forms/Constants.kt
@@ -1,4 +1,4 @@
-package com.klaviyo.messaging
+package com.klaviyo.forms
 
 import org.json.JSONArray
