+++ sdk/forms/src/main/java/com/klaviyo/forms/KlaviyoWebView.kt
@@ -1,4 +1,4 @@
-package com.klaviyo.messaging
+package com.klaviyo.forms
 
 import android.annotation.SuppressLint
 import android.app.Activity
